requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==0.4.6
altair==4.2.2
antlr4-python3-runtime==4.8
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8#egg=clip
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
diffusers==0.12.1
einops==0.3.0
entrypoints==0.4
fancycompleter==0.9.1
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.52.0rc1
gTTS==2.3.1
huggingface-hub==0.12.0
humanfriendly==10.0
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==6.0.0
invisible-watermark==0.1.5
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.0
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.7.0rc1
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.1
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.13.0
onnxruntime==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pdbpp==0.10.3
Pillow==9.4.0
platformdirs==2.6.2
proglog==0.1.10
protobuf==3.20.3
pudb==2019.2
pyarrow==11.0.0
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pydeck==0.8.0
pyDeprecate==0.3.1
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrepl==0.9.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
semver==3.0.0.dev4
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.17.0
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2023.2.2
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0.dev20230204
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.15.0.dev20230203
tornado==6.2
tqdm==4.64.1
transformers==4.19.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
urwid==2.1.2
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.2.2
wmctrl==0.4
yarl==1.8.2
zipp==3.12.0
